Added password requirements and finished item mongo implementation

+++ src/main/java/com/congapp/ShelterInv/api/DonatorController.java
@@ -1,9 +1,7 @@
 package com.congapp.ShelterInv.api;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,11 +13,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.yaml.snakeyaml.error.Mark;
 
 import com.congapp.ShelterInv.model.Donator;
 import com.congapp.ShelterInv.service.DonatorService;
-import com.congapp.ShelterInv.service.ShelterService;
 
 @RequestMapping("api/v1/customer")
 @RestController
@@ -49,13 +45,13 @@ public class DonatorController {
     }
 
     @DeleteMapping(path = "{id}")//User PSW
-    public void removePerson(@PathVariable String id){
-        personService.removePerson(id);
+    public void removePerson(@PathVariable String id, @RequestBody String password){
+        if (password.equals(personService.getPassword(id))) personService.removePerson(id);
     }
 
     @PutMapping(path = "{id}")//User PSW
-    public void changePersonName(@PathVariable String id, @RequestParam String name){
-        personService.changePersonName(id, name);
+    public void changePersonName(@PathVariable String id, @RequestParam String name, @RequestBody String password){
+        if (password.equals(personService.getPassword(id))) personService.changePersonName(id, name);
     }
 
 }
